Move *Extractor object instantiation earlier.
Start moving towards earlier instantiation of *Extractor objects

Rather than passing class names around, just instantiate the object
there and then.

Delete the Extractor class, the factory functionality is no longer
required and the initialization methods for the previous subclasses
ProcExtractor and AutoExtractor are diverging.

The only requirement is that they have an `extract` method and we will
worry about how to enforce that later.